Pure refactor to use flake-parts
Specifically, [the "every file is a flake-parts module" pattern](https://discourse.nixos.org/t/pattern-every-file-is-a-flake-parts-module/61271).

This is a truly pure refactor, in the sense that the derivations
produced by the flake at this rev are identical to the ones produced by this
flake at the parent rev.

In order to achieve this refactor-purity, the following were used and
can be removed at your convenience:

- `lib.mkBefore`
- `lib.mkAfter`
- `lib.mkOrder`

The reason I insisted on this refactor-purity is to make it easier for
you to review. Since there are no changes to the derivations, you have
much less to consider while reviewing.

However, some checks were added to the flake, for the benefit of
enjoying `nix flake check`.

This PR might not be perfect, but it's 01:43 here and I've been working
on it since ~10:40 of the previous day. So I will submit it as is.

Throughout, some `# SUGGESTION` comments are sprinkled for your
consideration.

It also adds formatting to the flake, which can be run using `nix fmt`
and also includes a formatting check that is included in `nix flake
check` and also in git hooks that are set up automatically if you enter
a dev shell using `nix develop` or even better
[direnv](https://direnv.net/) ([home-manager module](https://nix-community.github.io/home-manager/options.xhtml#opt-programs.direnv.enable)).

If you need any help reviewing, I'm available async and sync. But I live
in Thailand so sync mileage may vary.